+++ beaker/services/account.py
@@ -3,6 +3,7 @@ from typing import List
 from cachetools import TTLCache, cached
 
 from ..data_model import *
+from ..exceptions import *
 from .service_client import ServiceClient
 
 
@@ -26,3 +27,19 @@ class AccountClient(ServiceClient):
         List all organizations you are a member of.
         """
         return [Organization.from_json(d) for d in self.request("user/orgs").json()["data"]]
+
+    def get(self, account: str) -> Account:
+        """
+        Get information about an account.
+
+        :param account: The account name or ID.
+
+        :raises AccountNotFound: If the account doesn't exist.
+        """
+        return Account.from_json(
+            self.request(
+                f"users/{self._url_quote(account)}",
+                method="GET",
+                exceptions_for_status={404: AccountNotFound(account)},
+            ).json()
+        )