+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Beaker.experiment.tasks()`.
 - Added `Beaker.experiment.results()`.
 - Added `Beaker.experiment.spec()`.
+- Added `Beaker.workspace.datasets()`.
+- Added `Beaker.workspace.experiments()`.
+- Added `Beaker.workspace.images()`.
+- Added `Beaker.workspace.list()`.
+- Added `Beaker.account.get()`.
+
+### Removed
+
+- Removed `Beaker.experiment.list()`. Please use `Beaker.workspace.experiments()` instead.
 
 ## [v0.5.5](https://github.com/allenai/beaker-py/releases/tag/v0.5.5) - 2022-04-10